requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pillow== 11.3.0
requests